Clean up smells/list-modify-iterating check
The existing implementation of this check had to use a large number of
patterns, as matching class inclusion in a pattern means only
a statement can be matched.

Instead, use the pattern-inside helper to determine class inclusion, in
which case a pattern can match any expression.

As a bonus, the check now reports the smell line, as opposed to the line
at which the enclosing class is defined.