Copy secure-set-cookie to django
This was a relatively successful check in our flask corpus. I've ported
it to Django here.

Interestingly, there's not a good way to identify that the set_cookie
method is being called on an HttpResponse object without some
rudimentary typing or dataflow. So I just use the common identifier
"response" as the invocation target. Note that this means that this
check fires on Flask projects.

In order to have this not fire on random Django middleware vendored
code, I excluded expressions from common Django middlewares.

After doing this, the fire frequency on non-test, non-vendored code is
1 finding in the entire Django-500 corpus. It's a true positive.

For details, see:
https://dev.massive.ret2.co/jobs/618?page=1&sort_on=repo_url&sort_by=asc