Add python mutable default datastructure checks (#209)
* Add python mutable default datastructure checks

In python, function definitions with parameters set to have a default
value of `[]` or `{}` actually have this datastructure instantiated at
start up time and all calls to the function with the default value used
all point to the same location in memory as the startup instantiated object.

This usually results in unintuitive behavior of changes to the default
object persisting over multiple calls to the function. i.e.:

```
def f(x=[]):
	x.append(1)
	print(x)

f()  # [1]
f()  # [1, 1]
```

This adds test cases of this phenomenon and instances when the mutable
structure is appropriately handled/copied. Also does a first pass effort
to include some patterns for sgrep to find instances of default datastructures
being mutated.

* fixup! Add python mutable default datastructure checks

* fixup! Add python mutable default datastructure checks