Remove FP in python/smells/list-modify-iterating
Leverages deep-expression matching to remove a false positive where this
would fire on _any_ list modification in a for loop, even if it was not
the list under iteration.

A downside of this change is that the rule now fires on the enclosing
iteration, instead of the offending list modification.