Fix regression using latest semgrep
I've recently added ';' in the generic AST, but this introduced
some regressions in semgrep-rules because the range changed
and if you're using pattern-not or pattern-inside, this has implications.
Indeed if one pattern use ';' and the other not, then the range are
different and the intersection or negation might not trigger.

I didn't anticipate this problem ... It might be better to remove
back those ';' in the generic AST, but at the same time it also
allows better autofixing in the futur and better matching display
(we will highlight the ';' in semgrep.live if your pattern uses one).

Test plan:
make test (with latest semgrep-core locally;
hopefully the docker image will catch up)