Detect @Html.Raw on JSON
A common pattern in ASP.NET is to convert a backend C# object to JavaScript,
using the following pattern in the Razor template:

```
<script>
var a = @Html.Raw(Json.Encode(Model));
</script>
```

This encodes `Model` to JSON and assigns it to `a`. The `Html.Raw` is used to
avoid HTML encoding. Unfortunately, this also makes it vulnerable to XSS. By
including `</script>` in the JSON, it is possible to break out of the script
context.

This rule detects this pattern, the combination of `@Html.Raw` with JSON
encoding. I chose not to match the surrounding `<script>`. This makes it more
specific where the error occurs, and this pattern would also be vulnerable
outside of JavaScript.