Add rule for structured logging in C#
String interpolation should be done by the logging library instead of by
$"...", so that querying variables is possible, and log injection is prevented.

Add rules for various logging libraries. Some methods are double (e.g.
$LOG.Debug), since they appear in multiple libraries. I have kept this
intentionally to make it clear which methods belong in which libraries.

Also check that the method is called on something that is named "log" or
something like that. Logger objects could be named differently, but this
reduces the chance of false positives when calling `Trace` or something on an
object that is not a logger.