Add more acceptable cases for python requests use-raise-for-status rule
Sometimes the actual request will be made inside of a context manager,
but you don't want to check the status until after the context manager
is closed. For example, your context manager may have resulted in
opening a file or a different network socket, and closing it ASAP is
desirable. You're still checking the status of the request though, which
should be fine.