Allow a template variable in the nonce attribute of a script tag.
On receiving a CSP, a compliant browser will refuse to execute inline
scripts unless the server tags them with a nonce or hash.

The django-csp library handles generation and insertion of nonces. It
requires however that we indicate the point in which it should place
the nonce, using a template variable.

This PR modifies the var-in-script-src rule so that it will not flag a
template variable inside a nonce attribute.

References:
https://content-security-policy.com/script-src/
https://django-csp.readthedocs.io/en/latest/nonce.html