Fix dompurify rule autofix
The pattern targets an expression, but the fix was a statement, since it
is terminated by a semicolon. Removing the semicolon in the fix makes it
consistent.

I've added a test case that illustrates the bad behavior that this leads
to. Before this change, the new call was fixed to the following:

```
dosomethingsketchy(DOMPurify.sanitize(dirty, {RETURN_DOM: true}););
```

Note that this has an extra semicolon before the last close parenthesis.