Fix perf of ruby.lang.security
In principle, for these rules, this:

```
...
def $METHOD
  ...
end
...
```

should be equivalent to just `...`, but the latter is much faster.

(Yet I'm not sure why the current form is so slow...)

test plan:
% semgrep -c hardcoded-secret-rsa-passphrase.yaml path/to/gitlabhq/app/models/ci/build.rb
Before the change, on a MacBook Pro (Apple M1), this took ~3s, after the change, ~0.1s.
% semgrep -c insufficient-rsa-key-size.yaml path/to/gitlabhq/app/models/ci/build.rb
Before the change, on a MacBook Pro (Apple M1), this took ~20s, after the change, ~0.1s.