Update `server-dangerous-object-deserialization` rule
I'm making a change to the Semgrep Pro Engine which will correctly model
the fact that `java.lang.Object` is at the top of any inheritance
hierarchy in Java. As such, the pattern `Object $X` will match any
object type, including the `String` types used in the tests here.

In general, we do want the Pro Engine, when presented with a pattern
`(Foo $X)`, to also match any subtypes of `Foo`.

Based on a discussion with Pieter, this should actually match any object
type except for `String`s and boxed types. As such, I have updated this
rule and the test cases accordingly. Now, it functions the same both on
the Pro Engine and OSS, and is more accurate than before on both.