hashids-with-django-secret
The Django secret key is used as salt in HashIDs. The HashID mechanism is not secure.
By observing sufficient HashIDs, the salt used to construct them can be recovered.
This means the Django secret key can be obtained by attackers, through the HashIDs.

This is a rare bug, but easy to detect.