hashids-with-django-secret (#3161)
* hashids-with-django-secret

The Django secret key is used as salt in HashIDs. The HashID mechanism is not secure.
By observing sufficient HashIDs, the salt used to construct them can be recovered.
This means the Django secret key can be obtained by attackers, through the HashIDs.

This is a rare bug, but easy to detect.

* hashids-with-flask-secret

The Flask secret key is used as salt in HashIDs. The HashID mechanism is not secure.
By observing sufficient HashIDs, the salt used to construct them can be recovered.
This means the Flask secret key can be obtained by attackers, through the HashIDs.

* Add subcategory vuln

Linter complained about subcategory missing.
This seems like an actual problem that should be resolved.

---------

Co-authored-by: Pieter De Cremer (Semgrep) <pieter@r2c.dev>