Remove incorrect fix for cookie-issecure-false.yaml
This fix is seemingly copied from another rule. This rule checks for missing `setSecure(true)`, but the fix replaces a `setSecure(false)` with a `setSecure(true)`.

For a good autofix, this rule needs to be split into two rules:
- One for a missing `setSecure` call, that adds one with the right arguments
- One for a `setSecure` call with `false` that needs to be replaced with `true`.

Neither will need `fix-regex`, but can be fixed with a regular `fix`.